Added `[p]restart`
Makes Red quit with exit code 26. This makes the launcher able to catch the request and restart Red on demand.
The method that this command calls has been integrated into the `Bot`class so that cogs can easily call it as well, if needed.
Both [p]shutdown and [p]restart now both respond in chat when issues but can be optionally silenced by passing a True bool (which, in discord.py, equals to "true", "yes", etc...)