Fix pagify bug
Right now it seems pagify() will yield an empty string if it can't find a deliminator, and on looking at the code it looks like it's because it finds the deliminator at the start of the string if it's the second page or higher (since it was just cut off there for the previous string). By replacing the second rfind argument with len(d), I think that'll fix that, and limited testing shows that it does.