Fixes an issue with patch notes not showing
There was a discovery that the way repos are being added could lead to
multiple path separators being used within the "file" and "folder"
strings. This was brought about by "/" being used staticly in some
loctions of the downloader and the use of "os.path.join" in others.
However, it while the windows operating system, can handle this, the git
program installed cannot.
Until one path delimters method is settled on and appiled, I've
temporarily resolved the patch note issue by recreating the repo
location, and the cogfile within it using "os.path.join" so git now get
the correct values.