added async support for io thru aiofiles
also added backwards compatibility so if this is updated and deps aren't nothing will break. I could add in a forceful update of deps but I dont want to control the local environment of the end user. This is very hastily written though so highly advise it to be reviewed and tested