[V3 Core] Allow central storage of API keys
This creates a central location to store external API tokens that can be used between cogs without requiring each cog to be loaded for it to work.

A new set option for `[p]set api` is created to assist in forming bot readable api token locations.

This also updates the Streams cog to utilize the central database.