Fix grammatical errors.
In the sentence "mm:ss until track playback: #X in queue", it makes more
sense to have a comma instead of a colon, as the "#X in queue" part is
not dependent on the former part of the sentence. Also added a full
stop at the end of the sentence.