If empty dc enabled, check for vc participants 1st
When the bot is starting up and starting autoplay players, if empty dc is enabled on the server, check the vc first before reconnecting. If there is no one in the channel to listen, don't connect, and set the guild's autoplay config for that server to no channel so it won't reconnect. Otherwise the bot would connect on bot start, wait the timer out for the empty disconnect timer, and then disconnect anyway.