New ClearChannel command
Since deleting messages is limited to 14 days, here is the "[p]cleanup clearchannel" command to dupe a channel and delete the old one in order to remove all messages from it. The request must be approved by a "y" message like the other one and the "manage_channel" permission.
Thanks in advance.