feat: Config.get_conf annotation overloading
This commit add overloading to the `Config.get_conf` method.
The goal is to indicate that either `cog_instance` or `cog_name` must be provided.
Before, giving `None` to both arguments would not raise an error.
`cog_instance` now indicate that it requires a `discord.ext.commands.cog.Cog` instance

Last, but not least, removing `Any` from yield return of `_ValueCtxManager.__await__` because nothing is yielded...