Docs - More external -> unmanaged
Kreusada and I had discussed this change earlier when we were working on the docs together but I wanted to leave some mentions of "external" for users to recognize and be familiar with previously. I've changed my mind on this a bit and tried to reword some things like definitions of "unmanaged" and "managed" in some regards to make it hopefully, a clearer change for the users when we get there.