Add `--cog-path` CLI argument
This argument takes a list of strings (can be used multiple times for
more than one value), validates that the path provided exists, and adds
the path via `self._cog_mgr.add_path()`, so that cogs can be loaded from
that path.

This also adds the `persist` argument to `add_path()`, to allow for this
functionality.

Closes #6506